fix compile and test errors on scala3
Fixes some minor issues that cause compile errors on
scala3 for atlas-core. The atlas-core project now
compiles and all tests pass except for AlgoStateSuite
due to issues with atlas-json.

+++ atlas-core/src/test/scala/com/netflix/atlas/core/util/RefIntHashMapSuite.scala
@@ -147,7 +147,7 @@ class RefIntHashMapSuite extends FunSuite {
     (0 until 10000).foreach { i =>
       val v = Random.nextInt()
       imap.put(v.toLong, i)
-      jmap.put(v, i)
+      jmap.put(v.toLong, i)
     }
     assertEquals(jmap.toMap, imap.toMap)
     assertEquals(jmap.size, imap.size)
